==> 01-revisao.qmd <==
# Revisão Bibliográfica

## Suspensão de Plataformas no Brasil

No século XXI, o capitalismo está centrado na extração e uso de um tipo de material em particular: dados, cuja recordação se torna cada vez mais acessível por tecnologias que digitalizam atividades e dão lugar para novas indústrias centradas em plataformas digitais [@Srnicek2017]. Segundo @Dijck2019, plataformas digitais são arquiteturas programáticas desenhadas para organizar interações entre usuários, alimentadas por dados, automatizadas e organizadas por algoritmos e interfaces, formalizadas através de relações de posse impulsionadas por modelos de negócios e regidas por termos de uso. Sobre o papel das plataformas digitais nos mercados emergentes de extração de dados, é possível observar que:

> Ao fornecer um espaço digital para que outras pessoas interajam, as plataformas se posicionam de forma a extrair dados de processos naturais (condições climáticas, ciclos de cultivo etc.), de processos de produção (linhas de montagem, fabricação em fluxo contínuo etc.) e de outras empresas e usuários (rastreamento na web, dados de uso etc.). Elas são um aparato extrativo de dados [@Srnicek2017, p. 32 tradução nossa].

A adoção em larga escala de plataformas é capaz de moldar as interações e causar disrupções societais, ao mesmo tempo que estas são moldadas pelo ecossistema em que ganham forma [@Dijck2019]. O potencial de influência dessas é evidente em acontecimentos como a adoção de dados de usuários do Facebook pela Cambridge Analytica para influenciar a eleição de 2016 dos Estados Unidos, bem como a disseminação de notícias falsas que o Projeto de Lei 2630/2020, conhecido por "PL das *Fake News*", busca combater [@Pereira2024]. 

Em resposta à crescente influência das plataformas, os Estados Nacionais buscam tomar medidas para proteger sua soberania nacional da influência dessas *Big Techs* globais [@Pereira2024], com as legislações europeia e francesa buscando definir uma categoria específica para as plataformas digitais desde 2015 [@Seto2021]. @Urupa2020 destacam como sendo motivadores a regulação de plataformas digitais três fatores principais: o crescimento do poder desses agentes, os problemas envolvendo a relação assimétrica com usuários e a insuficiência dos instrumentos normativos para lidar com problemas, garantir competição e assegurar direitos e deveres legais. Nesse contexto, um dos principais desafios para a regulação é o fato de estados competirem entre si por controle sobre o meio digital com seus poderes assimétricos, de forma que, embora tenham a capacidade de determinar a natureza e velocidade da mudança, as organizações globais mantêm o poder de controlar a direção da mudança através das plataformas digitais [@Floridi2019].  Em função disso, as Nações Unidas lançaram em 2019 o “Painel de Alto Nível para a Cooperação Digital”, que inclui recomendações como o asseguramento de bens públicos, o combate às violações de direitos humanos e o estabelecimento de um sistema de governança que inclua também usuários e não seja formado apenas por agentes privados (como as plataformas) [@Urupa2020].

No contexto brasileiro, a Lei do Marco Civil da *Internet*, que entrou em vigor em 2014, surge da necessidade de definir os fundamentos de disciplina do uso da internet após uma intensa revolução digital, fundamentando-se em três pilares: neutralidade da rede, privacidade de usuários e liberdade de expressão do pensamento [@Leite2016]. Ainda assim, @Barbosa2023 descreve a regulação de casos envolvendo plataformas digitais no Brasil como mais conservadora em comparação às medidas regulatórias digitais tomadas pela União Europeia, cujas normas buscam extensivamente proibir práticas abusivas e proteger a competitividade. Similarmente, @Seto2021 destaca através da revisão de literatura a possibilidade pouco discutida de o Estado oferecer alternativas públicas ao serviço de plataformas privadas e incentivar iniciativas cidadãs, em vez de somente buscar regular a atuação das existentes.

No Brasil, casos de regulação de plataformas digitais são esparsos, porém de grandes proporções: em janeiro de 2007, a Justiça de São Paulo bloqueou o acesso ao YouTube por 48 horas após um processo movido por Renato Malzoni, que pedia a remoção permanente de um vídeo com cenas íntimas entre ele e a apresentadora de TV Daniela Cicarelli, filmados em uma praia da Espanha por um *paparazzi* e postado na plataforma sem autorização do casal [@Folha2007]. Similarmente, em outubro de 2013, um juiz de São Paulo intimou a bloquear o acesso ao Facebook em todo o país se a plataforma não removesse publicações consideradas ofensivas da modelo e apresentadora de TV Luize Altenhofen, porém o Facebook cumpriu a ordem e o bloqueio não ocorreu [@Estadao2013]. Em dezembro de 2015, o aplicativo de mensagens WhatsApp foi suspenso por 12 horas após decisão judiciária da juíza Sandra Regina Nostre Marques, da 1ª Vara Criminal de São Bernardo do Campo, em razão da plataforma não ter atendido a uma ordem judicial que solicitava dados referentes à uma investigação do Primeiro Comando da Capital (PCC) [@Estadao2015]. Analogamente, o Telegram foi bloqueado nacionalmente por 48 horas em março de 2022 por descumprir determinações judiciais, cujas exigências incluíam a nomeação de um representante legal no Brasil, a apresentação de um plano de combate à desinformação e a remoção de links do canal oficial de Jair Messias Bolsonaro que davam acesso a documentos de um inquérito sigiloso. O cumprimento dessas ordens pelo fundador do aplicativo, Pavel Durov, levou ao fim imediato da suspensão [@G12022]. Nestes casos, o Marco Civil da *Internet* que a defende como sendo "essencial ao exercício da cidadania" [@Brasil2014] é colocado em questão pelo poder do Estado de realizar o desligamento repentino de aplicativos ou plataformas digitais, poder esse que pode ser visto como equivalente ao que @Wu2010 descreve como "*master switch*" ou chave mestra da comunicação, um alerta sobre o perigo da existência de tamanho poder para as democracias. 

O caso mais recente de suspensão de plataformas digitais foi o bloqueio da rede social X (antigo Twitter), em função do descumprimento de ordens judiciais pela nomeação de um representante legal no Brasil [@G12024]. O bloqueio durou cerca de 5 semanas, uma das suspensões mais longas de plataformas digitais no país. Esse acontecimento, suas consequências e relevância serão abordados com maior aprofundamento no decorrer da redação.



## Marketing de Redes Sociais

De acordo com @Kaplan2010, redes sociais são um grupo de aplicativos hospedados na Internet, que se baseiam nos fundamentos ideológicos e tecnológicos da Web 2.0 e que permitem a criação e o intercâmbio de conteúdo gerado pelos usuários. Dessa forma, as redes sociais virtuais se destacam como um mecanismo de comunicação que permite um maior alcance de informações e interações do que aqueles possíveis em canais anteriores à era digital, mudando a forma com a qual pessoas interagem entre si e com marcas. Em particular, a natureza interativa desses canais transformou os consumidores de observadores passivos de conteúdo em participantes ativos que agora criam grandes quantidades de conteúdo por meio de suas conversas, interações e comportamentos on-line [@Malthouse2013].

O termo rede social é abrangente em definição, reunindo ambientes virtuais com funcionalidades muito diferentes entre si. Em vista disso, os autores @Kaplan2010 definem seis tipos de redes sociais de acordo com a forma em que os processos sociais ocorrem nestas: projetos colaborativos (como Wikipedia), blogs e microblogues (como X, antigo Twitter), comunidades de conteúdo (como YouTube e Twitch), sites de networking social (como Facebook e LinkedIn), mundos de jogos virtuais e mundos sociais virtuais (como Second Life). Redes sociais que surgiram depois de 2010, como TikTok e Instagram, ainda podem ser categorizadas dentre os seis tipos estabelecidos pelos autores como comunidade de conteúdo e site de networking social, respectivamente, de acordo com suas estruturas. Dessas, as mais comumente referidas pelo termo "redes sociais" são os sites de networking social e as plataformas de microblogs, com redes sociais como X e BlueSky unindo características de ambas as categorias em suas funcionalidades.
  
O acesso rápido e amplo aos consumidores-alvo, até mesmo em nível internacional, é uma vantagem das redes sociais para esforços de marketing [@Ravi2021], com campanhas promocionais digitais estando entre as estratégias mais comuns para agências e organizações de publicidade ou comunicação. Essas estratégias, conhecidas como *Social Media Marketing* (SMM), têm como principal atrativo o retorno rápido ao mesmo tempo que permitem informar consumidores da marca, construir posicionamento e adquirir *feedback* a custos mais baixos que através de canais físicos.
  
Com a instalação das redes sociais no cotidiano dos consumidores, @Tuten2015 alegam que um novo “P” pode ser adicionado ao mix de marketing, resultando nos 5 “P”s de praça, produto, promoção, preço e participação em marketing de redes sociais:
  
>Os canais de redes sociais oferecem aos consumidores um acesso sem precedentes. Os consumidores discutem, contribuem, colaboram e compartilham — com as marcas e entre si. A cultura do marketing mudou para uma cultura informal, focada na crença de que os clientes estão no controle. O guru do marketing Peter Drucker disse uma vez a famosa frase: “O objetivo de uma empresa é criar um cliente”. Com o alcance e a influência da comunidade nas mídias sociais, podemos expandir essa definição: o objetivo de uma empresa é criar clientes que criem outros clientes. Essa participação no processo é o novo quinto P do marketing (op.cit., p. 21, tradução nossa).



### Influenciadores digitais ou *Social Media Influencers*

Em explorar como as relações sociais são construídas em ambientes virtuais, @Baym2010 descrevem como um espaço (plataforma) e práticas (linguagem, hábitos, nacionalidade ou interesses) em comum contribuem para a formação de comunidades em redes sociais, que desenvolvem normas próprias e as perpetuam através da interação entre membros. Sites de networking social como o Facebook tendem a incentivar o surgimento de comunidades na incorporação de ferramentas para auto-apresentação, que fornecem categorias com as quais criar identidades, incluindo dados demográficos como idade, local de residência e interesses gerais. Tais ferramentas também podem ter o resultado não-intencional de facilitar a construção de identidades falsas ou alteradas, levando a um entendimento geral construído ao longo do tempo da distinção entre personalidades *on-line* e *off-line* [@Baym2010].

A honestidade percebida na forma em que uma personalidade virtual se porta em relação a sua comunidade é crucial para sua reputação e influência, fazendo com que outros usuários a vejam como semelhantes a eles e considerem mais valiosas suas opiniões ou recomendações [@Forbes2016]. Esse princípio, sustentado pela teoria de aprendizagem social, é o que dá origem aos influenciadores de redes sociais (social media influencer - SMI), representando um novo tipo de endossantes independentes que moldam as atitudes do público por meio de blogs, postagens e do uso de outras mídias sociais [@Freberg2011]. Atuando como pseudo-celebridades altamente segmentadas, estes influenciadores digitais representam para o marketing de influência uma ferramenta de relações públicas bem-sucedida que permite que as marcas e os consumidores se conectem em um nível mais pessoal [@Glucksman2017]. 

Os relacionamentos parassociais, inicialmente conceituados por @Horton1956, descrevem vínculos unilaterais com figuras da mídia. Na atualidade e através das redes sociais, esses relacionamentos são cada vez mais comuns, com o número de seguidores e a autoridade percebida dessas personalidades ampliando o impacto das mensagens emocionais propagadas por estes [@Giles2002]. Outros aspectos, como conhecimento e experiência, auto-confiança, articulação na fala e grau de identificação com o consumidor-alvo demonstrados são características importantes de um influenciador, que afetam a confiabilidade de seu *word-of-mouth* (marketing boca-a-boca) eletrônico [@Forbes2016]. Sendo assim, a multitude de atributos de um influenciador digital, moldados ou não a partir de sua identidade verdadeira, afetam a eficácia do marketing de influência proporcionalmente à sua compatibilidade com a imagem que a marca deseja propagar, juntamente de características gerais que estão diretamente relacionadas à capacidade de influenciar.



## Indústria de *Games* enquanto Indústria Criativa

Jogos eletrônicos ou *games*, além de servirem como passa-tempo e atividade de lazer, representam um ambiente de dados empíricos valiosos para a criação de conhecimento, exercendo papel central na mídia, redes e cenário sociotécnico da atualidade [@Taylor2019]. Segundo dados do 12º Relatório PGB [@PGB2025], desenvolvido pela Go Gamers e SX Group em parceria com a ESPM e Blend New Research e aplicada nos países Argentina, Chile, Colômbia, México, Peru e Brasil, 82,8% dos participantes relataram ter o costume de jogar algum tipo de jogo digital, um crescimento de 8,9% em relação à edição anterior (73,9%), de 2024.

@Terranova2000 incentiva o debate sobre como trabalhos culturais e técnicos da internet são uma característica predominante da economia pós-industrial, em que indivíduos contribuem com esforços laborais para plataformas comercializadas desde a produção de vídeos no YouTube à criação de postagens em páginas pessoais de redes de networking social. No contexto de jogos eletrônicos, essas plataformas ofereceram algumas das primeiras oportunidades e normas para a monetização do ato de jogar, bem como de atividades criativas na criação de conteúdo através de jogos eletrônicos, amplamente descrito por *user generated content* (UGC) [@Taylor2019].

A transmissão de jogos eletrônicos como uma atividade monetizada surge de uma multitude de fatores. Para aquele que realiza as transmissões, o *streamer*, os principais motivadores são as conexões sociais (como o desejo de compartilhar o ato de jogar com amigos), a amplificação da experiência de jogo através da introdução de telespectadores, as performances criativas e capacidade de auto-expressão através da interação com uma audiência expressiva e as aspirações profissionais de sustentar-se através de um *hobby* ou contribuir para o crescimento indústria de *games* através da publicidade [@Taylor2019]. Em especial, o aspecto relacional das transmissões ao vivo é explicado por @Taylor2019 como um componente importante para o aspecto financeiro da atividade, onde doações e inscrições são apresentadas como formas de expressar gratidão ao criador de conteúdo e tendem a ser seguidas por agradecimentos diretos deste. Desta forma, *streamers* atuam de forma semelhante à influenciadores digitais e possuem um incentivo financeiro direto para promover a parassociabilidade de sua audiência, cuja generosidade é a principal fonte de remuneração.



## Indústria de *E-Sports*

Existe um extenso debate sobre a diferenciação ou a semelhança dos esportes eletrônicos com os esportes tradicionais [@Wollesen2025; @Block2021]. O presente trabalho não busca expandir a discussão da classificação ou não dos esportes eletrônicos como esportes. Pelo contrário, ele já carrega essa premissa e busca expandir os horizontes da aplicação do cenário a partir disso. Utiliza-se então da definição dos autores @Juho2017, em que *e-sports* referem-se a partidas competitivas (amadoras ou profissionais) de *videogames* que podem ser coordenadas por instituições que realizam torneios, onde os *players* (jogadores) ou *pro-players* (jogadores profissionais) costumeiramente pertencem a times que podem ser patrocinados por outras instituições, gerando assim, um negócio.

Transmissões *on-line* (*streaming*) e eventos híbridos (campeonatos com torcida presencial que são também transmitidos de forma *on-line*) são as principais formas de disseminação dos *e-sports*, nos quais toda a atividade esportiva (antes, durante e após o evento) é mediada por computadores, facilitando a obtenção de dados dos telespectadores [@Juho2017]. Nesse contexto, @Cheung2011 definem o espectador de jogos eletrônicos como uma pessoa cujas intenções variam entre assistir ao jogo casualmente e ser um fã em torneios competitivos de jogos, fazendo parte de um "ecossistema de espectadores" onde normas sociais são estabelecidas e experiências emocionais são compartilhadas. Os autores também destacam os "círculos de observação" como sendo representativos das contribuições performáticas dos comentadores e da plateia que acontecem ao redor do jogo, cuja narração e diálogo contribuem para a criação de uma experiência mais agradável para o espectador [@Cheung2011].

Em sua *Motivation Scale for Sport Consumption* (Escala de Motivação de Consumo de Esportes), @Trail2001 identificam conquistas (sentir-se realizado com o sucesso do time), escapismo (uma fuga da vida comum), interação social, aquisição de conhecimento, habilidades físicas dos jogadores e atração física como os construtos motivacionais mais positivamente correlacionados ao consumo de mídias esportivas tradicionais e aos fãs de esportes tradicionais. Aplicada ao consumo de *e-sports*, muitos dos motivos permanecem os mesmos [@Cheung2011], porém a aquisição de conhecimento se destaca como o construto motivacional mais associado à intenção de assistir *e-sports* [@Macey2025].

No ambiente virtual das redes sociais, os torneios de *e-sports* são promovidos em grande parte por *mídias propagáveis* como memes, análises, clipes e reações [@Taylor2019]. De acordo com @Jenkins2015, propagabilidade se refere ao potencial cultural e técnico de públicos compartilharem conteúdos por motivos próprios, com ou sem a permissão dos detentores de direitos autorais. Aspectos que influenciam a propagabilidade de mídias incluem o fácil acesso, portabilidade, reutilizabilidade, relevância para públicos amplos e participação em um fluxo constante de material [@Jenkins2015]. A troca social inerente aos esportes, ampliada pelo fácil acesso à transmissão, reprodução e veiculação dos *e-sports* fazem deles mídias altamente propagáveis, em que fãs não são apenas consumidores ou telespectadores mas produtores ativos de conteúdo.



### League of Legends

Atualmente, League of Legends (LoL) se destaca como um dos jogos eletrônicos e *e-sports* mais relevantes do mercado, ocupando consistentemente a primeira ou segunda posição em jogos mais assistidos na plataforma Twitch no ano de 2025 [@TwitchMetrics]. O título, do gênero *Multiplayer Online Battle Arena* (MOBA, Arena de Batalha Multijogador *On-line*), é um jogo *on-line* desenvolvido pela empresa Riot Games em 2009, com as partidas nos modos de jogo principais sendo compostas por duas equipes de 5 jogadores, nas quais cada jogador toma controle de um personagem e são esperados a trabalharem em conjunto para destruir a base inimiga. [@Cauteruccio2023]

Devido à complexidade das partidas e o nível de habilidade demandado, um grande cenário competitivo vem crescendo ao redor do jogo ao passar dos anos. De acordo com @Cauteruccio2023, o principal evento de *e-sports* de League of Legends é o LoL World Championship (também conhecido como *Worlds*), onde as melhores equipes do mundo participam e disputam os campeonatos competindo por prêmios multimilionários em formato de torneio. As transmissões ao vivo das ligas geralmente são transmitidas por diferentes serviços de *streaming* como Twitch e YouTube e o número de visualizações são expressivos, com mais de 99 milhões contabilizadas mundialmente durante a final do Worlds de 2018 [@Cauteruccio2023].

Durante 2025, o cenário competitivo de League of Legends foi formado por cinco ligas regionais: Coreia do Sul (LCK), Américas (LTA), Europa, Oriente Médio e África (LEC), China (LPL) e Ásia-Pacífico (LCP), nas quais times de uma mesma região competem entre si por vagas nas competições internacionais [@LoL2025]. Anunciada em Junho de 2024 e fundada em 31 de outubro de 2024, a Liga de LoL das Américas (LTA) uniu as principais ligas americanas América do Norte – LCS, Campeonato Brasileiro de League of Legends – CBLOL e Liga Latinoamérica – LLA em um único ecossistema com duas conferências: LTA Norte (baseada na LCS) e LTA Sul (baseada no CBLOL), com ambas integrando ex-equipes da LLA [@Riot2024LTA]. No entanto, o fim da união LTA foi anunciado em menos de um ano após sua fundação, em 29 de Setembro de 2025, com a declaração do retorno do CBLOL e LCS como ligas independentes a partir de 2026, assegurando uma vaga no *Worlds* para o time brasileiro vencedor da liga regional [@CBLOL2025].



## Ciência de Dados

Com o avanço da internet, a coleta e a circulação de informações passaram a ocorrer em volumes e velocidades cada vez maiores. O valor destes dados são reconhecidos globalmente, com muitas organizações adotando estratégias orientadas por dados de forma a fazer decisões melhor informadas [@Wedel2016]. 

A necessidade de estudar dados como uma ciência à parte e não apenas através de suas aplicações em pesquisa e indústria leva ao surgimento da Ciência de Dados, que recebeu diversas definições ao longo dos anos. @Zhu2015 consolidam as várias conceituações ao definir Ciência de Dados como a teoria, o método e a tecnologia de estudar a natureza dos dados, tendo dois componentes principais. O primeiro componente é o estudo dos padrões e regras dos próprios dados, explorando a natureza dos dados sem necessariamente levar em consideração seu significado no mundo real. O segundo componente é o estudo das leis do mundo natural refletidas nos dados, de forma a inferir informações sobre a realidade através do estudo de dados.

A Ciência de Dados é por natureza um campo interdisciplinar, tendo aplicações comuns em diversas ciências pela forma que transforma dados desordenados em informações úteis que apoiam ou guiam decisões. De acordo com @Ozsu2024, o sucesso da ciência de dados é evidente em campos como as ciências da saúde, onde a análise de teias sociais permite o rastreamento de epidemias; em sistemas financeiros, em que a análise de grandes volumes de dados podem guiar investimentos e até mesmo para a indústria de atendimento ao consumidor, na qual o avanço de *chatterbots* auxilia no atendimento.

Para estudos de análise de marketing, a Ciência de Dados aplicada a contextos de Big Data (bem como em conjuntos de dados menores ou segmentados) é capaz de perfilar os clientes mais promissores de forma a melhor guiar esforços de propaganda através da análise dos históricos de navegação e hábitos *on-line* de uma grande quantidade de pessoas. Além disso, é possível identificar alternativas atraentes para um produto do ponto de vista do consumidor com embasamento estatístico, bem como refinar algoritmos de recomendação para melhor prever o que um consumidor estaria mais propenso a adquirir com base em suas compras anteriores. Através da análise de redes sociais e do monitoramento das conexões entre usuários, empresas são capazes de até mesmo avaliar associações de marca como o quão sustentável ela aparenta ser do ponto de vista de um consumidor médio, facilitando ações de marca reativas ou preditivas. [@Chintagunta2016]

De acordo com @Tabani2020, embora a análise de marketing seja um tópico altamente relevante na atualidade, o potencial dela em negócios reais ainda não é totalmente realizado em razão do crescente volume e complexidade de dados disponíveis através das redes sociais, ferramentas digitais e aplicações de software. O artigo aponta que barreiras na adoção desta abordagem incluem a falta de estratégias abrangentes, expertise, recursos e dados de qualidade, dentre outros fatores. @Hauser2007 descreve que para alcançar entendimentos práticos e aplicáveis para o marketing, é o papel do analista de dados minerar, analisar, interpretar e apresentar a informação de forma a convertê-la em inteligência acionável, de forma a acompanhar, segmentar e modelar o 'DNA informático' do consumidor. Neste contexto, é evidente que as informações em si não são tão úteis para organizações quanto a inteligência de mercado necessária para analisá-las. 

A análise de redes sociais têm emergido como uma ferramenta crítica para o entendimento do comportamento do consumidor na era digital, sendo capaz de providenciar discernimentos superiores aos de modelos tradicionais de segmentação demográfica e psicográfica [@Dam2022]. De forma abrangente, este método pode ser definido como toda forma de análise de dados estruturados (organizados em estrutura pré-definida) e não-estruturados (sem formato predeterminado) provenientes de redes sociais (*microblogues*, *sites de networking social* ou *comunidades de conteúdo*), independente do formato destas [@Gandomi2015]. Embora aplicações em marketing sejam a principal utilização da análise de redes sociais em tempos recentes, a diversidade de dados presentes nas redes sociais permite o uso em pesquisas em diversas disciplinas como psicologia, sociologia, antropologia e outras. 

Os autores @Gandomi2015 classificam as fontes de informações em redes sociais entre conteúdos gerados por usuários (sentimentos, imagens, vídeos e marcações) e interações entre usuários da rede (pessoas, organizações e produtos). Através desta divisão, os autores segmentam a análise de redes sociais em dois grupos: a análise baseada *em conteúdo* e a análise baseada *em estrutura*. A análise *baseada em conteúdo* tem como objeto de estudo os dados publicados por usuários, comumente desestruturados e ruidosos, através de métodos de análise de texto, áudio e vídeo como a análise de sentimento. Em contraste, a análise *baseada em estrutura*, busca analisar os atributos estruturais de uma rede social e extrair inteligência das relações entre as entidades participantes, como através de grafos sociais explicando conexões entre usuários.

Para a obtenção dos dados de redes sociais, a mineração é uma tática comumente utilizada para fins de pesquisa amplos. De acordo com @Barbier2011, uma definição de mineração de dados é a identificação de padrões novos e acionáveis nos dados, de forma a extrair novos conhecimentos de conjuntos de dados. Algoritmos supervisionados e não-supervisionados são utilizados para identificar padrões nos dados. Para algorítmos supervisionados, a *classificação* é um método em que um algorítmo é treinado com um conjunto exemplo de dados com rótulos para cada elemento, de forma a aprender a rotular dados semelhantes. Similarmente, *clustering* é um método de agrupamento para mineração não-supervisionada, que identificam semelhanças entre elementos em dados para determinar suas distâncias. Estes métodos de agrupamentos são semelhantes à *detecção de comunidade*, pela qual são extraídas conexões implícitas de uma rede [@Aggarwal2011]. 

@Gandomi2015 explicam a análise de influência social como outra técnica na análise de redes sociais, que agrupa métodos de modelação e avaliação da influência de usuários e conexões em uma rede social. Métodos de análise de influência social são amplamente úteis em esforços de *brand awareness* e adoção de marca, auxiliando no marketing de influência. Adicionalmente, os autores descrevem o método de *previsão de conexões* como a análise de dados passados para tentar prever interações futuras, de forma a levar em consideração a natureza dinâmica e imprevisível das dinâmicas de redes sociais.
