==> 1-Extracao.R <==
#1 - extrair postagens entre as duas datas
#2 - para cada postagem,
#2.1 - extrair todos os comentários (thread) (que serão novas postagens)
#2.2 - extrair reposts de todas as threads
#2.3 - extrair todos os likes de todas as threads
#2.4 - extrair todos os users de todas as threads

#3 - escrever em disco 3 dataframes (threads, likes e users)
