==> 02-probl-objet-justif.qmd <==
# Problema de pesquisa

Em dezembro de 2019, Jack Donsey, então CEO do Twitter Inc., anunciou que a empresa estaria financiando uma equipe de profissionais independentes para o desenvolvimento de um novo protocolo aberto e descentralizado para redes sociais, com a intenção de eventualmente implementá-lo para o próprio Twitter. Em 2021, o projeto se tornou uma organização independente do Twitter, dando origem ao Bluesky Social PCB. Um ano depois, em 2022, Elon Musk adquiriu o Twitter (renomeado como X) e a rede cessou todas relações de serviço com o Bluesky, que seguiu com sua proposta de desenvolver e impulsionar a adoção em larga escala de tecnologias para conversas públicas, abertas e descentralizadas [@Bluesky2023].

A rede teve um grande aumento de usuários em 30 de agosto de 2024, quando a *X Brasil Internet Ltda.* e outras empresas relacionadas à Elon Musk tiveram seus funcionamentos integralmente suspensos em todo o Brasil por determinação de Alexandre de Moraes, ministro do Supremo Tribunal Federal. O motivo da suspensão foi o desobedecimento de ordens judiciais brasileiras na não nomeação de um representante legal dentro do prazo estipulado pelo ministro na intimação (PET 12.404 DF) de 18 de setembro de 2024 [@G12024].

A suspensão do X no Brasil durou cerca de 40 dias, encerrando em 8 de outubro de 2024 após o cumprimento das ordens judiciais pela rede e uma multa de 28,6 milhões de reais. Durante esse período, a migração de usuários brasileiros para alternativas como o Bluesky foi impulsionada significativamente. Em 31 agosto de 2024, o Bluesky anunciou que havia batido mais de 1 milhão de novos usuários nos últimos 3 dias e que brasileiros estavam estabelecendo novos recordes de atividade na plataforma [@G12024].

O conceito de *sociedade plataformizada*, proposto por @Dijck2019, explica o fenômeno em que as lógicas das plataformas (como AirBnb, Google, Uber e outras) moldam a interação social, a economia e o exercicio da cidadania. Os autores dividem o ecossistema de plataformas entre plataformas infraestruturais e plataformas setoriais, em que as plataformas infraestruturais são as mais influentes e formam o coração do ecossistema sobre as quais outros aplicativos e plataformas setoriais podem ser construídos, como mecanismos de busca, navegadores, *e-mail*, plataformas de mensagem, redes sociais, lojas de aplicativos, *GPS*, sistemas de pagamento e hospedagem de arquivos. O segundo tipo, as plataformas setoriais, são aquelas que servem um nicho específico como alimentação, transporte, notícias ou hospitalidade [@Dijck2019]. Sendo assim, podemos entender a migração de usuários do X (centralizado, baseado em anúncios e de protocolo proprietário) para o Bluesky (descentralizado, sem anúncios e de protocolo aberto) como uma mudança de um ecossistema plataformizado fechado para um potencialmente mais aberto. Essa perspectiva é sustentada por @Plantin2018, que destaca como a utilização de interfaces de programação de aplicações (APIs) de plataformas infraestruturais por outros atores predominantes cria um ambiente virtual controlado por essas plataformas, que passam a "dominar um território" da internet anteriormente livre, centralizando-a para fins comerciais. O projeto que viria a se tornar o Bluesky surge em resposta a essa crescente centralização, buscando desenvolver um novo padrão descentralizado para conversas públicas em redes sociais (o Protocolo AT -- *Authenticated Transfer Protocol*).

Em seu artigo sobre as consequências da algoritmização na sociedade, @Velkova2021 abordam a ideia de "políticas de reparo" midiáticas como sendo trabalhos corretivos em conjunto com e dentro de sistemas algorítmicos para fazê-los gerarem resultados não intencionais, em resposta a tendenciosidades ou "falhas" percebidas nos resultados de algorítmos. Similarmente, a migração em massa de usuários brasileiros do X para o Bluesky pode ser vista como uma forma de "reparo" ou "resistência algorítmica" diante da falha (interdição no Brasil) de uma plataforma centralizada, tornando necessário o reparo de seus tecidos sociais e comunicativos em um novo espaço onde possam reconstruir comunidades.

Ao mesmo tempo em que o X foi suspenso ocorria a 13ª edição do Campeonato Brasileiro de League of Legends (CBLOL), um torneio competitivo de *e-sports* de League of Legends no Brasil organizado anualmente pela Riot Games. Em 2024, o evento teve sua primeira etapa entre 20 de janeiro e 20 de abril e sua segunda etapa entre 1 de junho e 7 de setembro. Com partidas gerando um grande número de engajamento nas redes sociais, a produção de conteúdo por espectadores durante a segunda etapa do CBLOL 2024 foi diretamente impactada pela suspensão da rede social no Brasil, que teve seu início próximo às finais do campeonato e término somente depois que ele havia encerrado. [@CBLOL2024]

Diante disso, levanta-se a seguinte questão central: de que maneira o bloqueio do X no Brasil influenciou as dinâmicas de engajamento e produção de conteúdo sobre *e-sports* de League of Legends na rede social Bluesky?



# Objetivos

## Objetivo Geral

- Caracterizar como a suspensão das atividades do X no Brasil influenciou as dinâmicas de engajamento e produção de conteúdo sobre *e-sports* de League of Legends na rede social Bluesky.

## Objetivos Específicos

- Mensurar a flutuação no volume de postagens no Bluesky sobre League of Legends antes, durante e após a suspensão do X no Brasil.
- Caracterizar as dinâmicas de seguidores e influência social de usuários dentro de comunidades associadas à League of Legends.

# Justificativa

No meio digital, esforços de pesquisa acadêmica devem considerar que os objetos de estudo são em grande parte instáveis, seja pela efemeridade de *websites* (que podem ser alterados, redirecionados ou retirados do ar subitamente), atualizações de *software* ou modificações em serviços digitais que reconfiguram de alguma forma aquilo sendo estudado [@Rogers2009]. Tal característica se torna evidente com o rebranding do Twitter para X e a reconfiguração subsequente da plataforma após a aquisição de Elon Musk, que encerrou o acesso gratuito à API e passou a cobrar um serviço de inscrição em Fevereiro de 2023 [@EconomicTimes2025]. Como consequência, a mineração ou raspagem de postagens para estudos de percepção pública, análise de sentimentos, predição de consumo [@Bhattacharjee2017; @Chakraborty2023; @Bian2016] e outras metodologias de pesquisa semelhantes que utilizassem de dados do antigo Twitter se tornaram repentinamente irreplicáveis.

Em função desta constante mudança, @Rogers2009 argumenta que através do emprego de métodos virtuais, pesquisas relacionadas à internet devem deixar de focar na mensuração de quanto cultura virtual reflete a realidade para, em vez disso, diagnosticar as mudanças culturais e societais *através* da internet, como fonte de dados e não objeto de estudo. A suspensão do X no Brasil e a subsequente migração de usuários para o BlueSky, portanto, é um fenômeno digital que reflete uma mudança cultural e social e que pode vir a ser estudada através de métodos digitais.

Como @Jenkins2015 descrevem, os contextos tecnológicos, econômicos e legais afetam a forma em que a participação cultural se manifesta, com o ambiente atual de mídia sendo cada vez mais propício à propagação de conteúdo pela facilidade de apropriação, remixagem, recirculação e incorporação na natureza das tecnologias. Através dos esforços laborais de usuários na criação de conteúdo e promoção, estas plataformas comercializadas se tornam mais atrativas para novos usuários que agem simultaneamente como consumidores e produtores [@Terranova2000]. Essas mídias propagáveis, portanto, são representativas de valores sociais ou culturais dos usuários envolvidos, como @Jenkins2015 explicam:

>[...] se um público vai tomar para si a propagação do conteúdo de mídia ou a discussão sobre aquele conteúdo, é porque esse conteúdo atende a algum
propósito comunicativo para o público e porque se encaixa em suas conversas que estão em andamento (op.cit. p. 467-468).

Um dos maiores contribuidores em volume de *user generated content* nas redes sociais é a comunidade *gamer*, cuja cultura de assistir outros jogarem, presente desde os fliperamas e consoles domésticos até as batalhas *on-line*, consolidou-se como parte essencial da experiência dos jogos [@Taylor2019]. Dentre os jogos eletrônicos mais relevantes atualmente, League of Legends se destaca como um *e-sports* dos mais assistidos [@Cauteruccio2023], nos quais fãs possuem investimentos afetivos nos jogadores e são capazes de compartilhar experiências emocionais dentro do ecossistema de espectadores [@Cheung2011]. Neste ambiente, os jogadores profissionais e *streamers* comentadores de *e-sports*, de atuação análoga aos influenciadores digitais, geram uma conexão mais próxima entre os espectadores e os jogos que transmitem [@Glucksman2017]. 
 
No Brasil, a 11ª edição do relatório da Pesquisa Games Brasil (2024) evidencia uma grande participação no mundo dos *e-sports*, com 81,2% dos respondentes relatando ter o costume de jogar jogos da categoria e 88,3% de assistir a partidas ou campeonatos relacionados [@PGB2024]. A união desses diversos fatores, pareados com a prominência dos jogos eletrônicos e dos *e-sports* na sociedade brasileira, fazem da discussão acerca de *e-sports* de League of Legends nas redes sociais um tópico com grande volume de dados digitais empíricos valiosos para a criação de conhecimento social. Portanto, a suspensão do X no Brasil, por coincidir com o CBLOL, torna possível examinar como as dinâmicas de engajamento e produção de conteúdo dessa comunidade no BlueSky foram afetadas pela migração através da análise de redes sociais e representa uma oportunidade de observar a construção da expressão *on-line* de uma rede de relacionamentos já existente, de forma a gerar uma melhor compreensão do impacto desse acontecimento digital. Ao mesmo tempo, representa também uma oportunidade de observar a construção da expressão *on-line* de uma rede de relacionamentos já existente.
