==> 2-tratamento-dados.qmd <==
---
title: "Tratamento dos dados de *likes* em *skeets*"
format: 
  html:
    toc: true
    toc-depth: 5
    toc-location: left
    embed-resources: true
execute: 
  cache: refresh
---

```{r}
#| label: setup
#| warning: false

library(tidyverse)
`%!in%` <- negate(`%in%`)
```

## Dados de likes

Montando lista de likes

```{r}
#| label: carrega-likes

likes <-
  list.files(
    path = "Dados/likes", 
    pattern = "*.rds",
    full.names = TRUE
  ) |> 
  map(
    \(nome) {
      read_rds(nome) |> 
        mutate(
          uri.skeet = 
            nome |> 
            str_remove("\\..*$") |> 
            str_replace("^", "at://did:plc:") |> 
            str_replace("-", "/app.bsky.feed.post/"),
          data.skeet = created_at,
          to.actor =
            nome |> 
            str_remove("\\..*$") |> 
            str_replace("Dados/likes/", "did:plc:") |> 
            str_remove("-.*$"),
          from.actor = 
            actor_data |> 
            map_chr(\(x) x$did),
          from.handle =
            actor_data |> 
            map_chr(\(x) x$handle),
          .keep = "none"
        )
    }
  ) |> 
  list_rbind() 

likes |> write_rds("Dados/likes.rds")
```

Criando listas de actors relacionados com likes

```{r}
#| label: criando-likes-actors

# quem curtiu
liker.actors <- 
  likes |> 
  select(from.actor, from.handle) |> 
  unique()
liker.actors |> write_rds("Dados/liker-actors.rds")

# quem foi curtido
liked.actors <- 
  likes$to.actor |> 
  unique()
liked.actors |> write_rds("Dados/liked-actors.rds")

# quem curtiu ou foi curtido
likes.actors <- 
  c(liker.actors$from.actor, liked.actors) |> 
  unique()
likes.actors |> write_rds("Dados/likes-actors.rds")
```

Montando a matriz de likes dados

```{r}
#| label: monta-matriz-likes

all.actors.likes <- 
  tibble(
    did = 
      c(likes$from.actor, likes$to.actor) |> 
      unique() |> 
      sort()
  ) |> 
  left_join(
    likes |> distinct(did = from.actor, handle = from.handle),
    by = "did"
  ) |>
  mutate(num = row_number()) |>
    mutate(handle = case_when(did == "did:plc:uk7niqvqdgr74qhwb4vrg3vi" ~ "LTA_SUL",
        TRUE ~ handle
      )
    )

all.actors.likes |> write_rds("Dados/all-actors-likes.rds")

matriz.likes <- 
  likes |> 
  left_join(
    all.actors.likes,
    by = c("from.actor" = "did")
  ) |> 
  rename(actor.influenced.num = num) |> 
  left_join(
    all.actors.likes,
    by = c("to.actor" = "did")
  ) |> 
  rename(actor.influencer.num = num) |> 
  select(
    data.skeet, 
    actor.influencer.num,
    actor.influenced.num
  )
matriz.likes |> write_rds("Dados/matriz-likes.rds")
```

## Dados de skeets

Carregando a versão curta

```{r}
#| label: carrega-skeets

skeets <- read_rds("Dados/skeets-curto.rds")
```

Montando lista de envolvidos em skeets

```{r}
#| label: criando-skeets-actors

all.actors.skeets <- 
  tibble(
    did =
      c(
        skeets$did.autor,
        skeets$uri.post,
        skeets$uri.reply,
        skeets$uri.quote
      ) |> 
      str_remove("^at://") |> 
      str_remove("/app.*$") |> 
      unique() |> 
      sort()
  ) |> 
  mutate(num = row_number())
all.actors.likes |> write_rds("Dados/all-actors-likes.rds")
  
matriz.skeets <- 
  skeets |> 
  # se nem é reply e nem é quote, não é uma interação
  filter(!is.na(uri.reply) | !is.na(uri.quote)) |>
  left_join(
    all.actors.skeets,
    by = c("did.autor" = "did")
  ) |> 
  rename(actor.influenced.num = num) |> 
  mutate(
    did.replied.to = 
      uri.reply |> 
      str_remove("at://") |> 
      str_remove("/app.*$"),
    did.quoted = 
      uri.quote |> 
      str_remove("at://") |> 
      str_remove("/app.*$")
  ) |> 
  left_join(
    all.actors.skeets,
    by = c("did.replied.to" = "did")
  ) |> 
  rename(num.replied = num) |> 
  left_join(
    all.actors.skeets,
    by = c("did.quoted" = "did")
  ) |> 
  rename(num.quoted = num) |> 
  mutate(
    actor.influencer.num = 
      case_when(
        !is.na(num.replied) ~ num.replied,
        !is.na(num.quoted) ~ num.quoted,
        .default = NA
      ),
    influence = 
      case_when(
        !is.na(num.replied) ~ "reply",
        !is.na(num.quoted) ~ "quote",
        .default = NA
      ),
  ) |> 
  select(
    data.post, 
    actor.influencer.num,
    actor.influenced.num,
    influence
  )
matriz.skeets |> write_rds("Dados/matriz-skeets.rds")
```

Separando os skeets com "cblol" em iniciais e replies/quotes

```{r}
#| label: separando-skeets-iniciais-replies-quotes

skeets.separados <- 
  skeets |> 
  mutate(
    tipo = 
      case_when(
        !is.na(uri.reply) ~ "reply",
        !is.na(uri.quote) ~ "quote",
        .default = "skeet"
      )
  ) |> # count(tipo)
  group_by(tipo) |>
  group_split(.keep = TRUE) |> # length()
  set_names(c("quote", "reply", "skeet"))
skeets.separados$quote |> write_rds("Dados/skeets-quote.rds")
skeets.separados$reply |> write_rds("Dados/skeets-reply.rds")
skeets.separados$skeet |> write_rds("Dados/skeets-iniciais.rds")
```

## Dicionário de actors

Precisa:

- juntar todos os did's já encontrados com os handles quando presentes
- numerar sequencialmente os "handle.invalid" de did's diferentes
- criar um "handle.missing" quando is.na(handle)
- remover o ".bsky.social" de quem tiver
- colar os handles dos did's que mudaram de handle pelo meio do caminho

Buscando os skeets longos

```{r}
#| label: carrega-skeets-longo

skeets.longo <- read_rds("Dados/skeets-longo.rds")
```

Montando a lista de did's e handles

```{r}
#| label: monta-did-handle

did.handle.unicos <- 
  bind_rows(
    skeets.longo |> select(did = uri, handle = author_handle),
    skeets.longo |> select(did = uri, handle = author_handle),
    skeets.longo |> select(did = in_reply_to),
    skeets.longo |> select(did = quotes),
  ) |> 
  filter(!is.na(did)) |> 
  mutate(
    did =
      did |> 
      str_remove("^at://") |> 
      str_remove("/app.*$")
  ) |> 
  bind_rows(
    likes |> select(did = from.actor, handle = from.handle),
    likes |> select(did = to.actor),
  ) |> 
  arrange(did, handle) |> 
  distinct()
```

Tem muita gente com múltiplos handles?

```{r}
did.handle.unicos |> 
  group_by(did) |> 
  summarise(
    quantos.handles = n(),
    any.handle.na = any(is.na(handle)),
    any.handle.invalid = any(handle == "handle.invalid")
  ) |> 
  count(
    quantos.handles,
    any.handle.na,
    any.handle.invalid,
  )
```

```{r}
did.handle.unicos |> 
  group_by(did) |> 
  filter((n() == 1) | (n() > 1 & !is.na(handle))) |> 
  summarise(
    quantos.handles = n(),
    any.handle.na = any(is.na(handle)),
    any.handle.invalid = any(handle == "handle.invalid")
  ) |> 
  count(
    quantos.handles,
    any.handle.na,
    any.handle.invalid,
  )
```

Não. Só há 3 situações:

- ou apenas handle NA, 
- ou apenas handle inválido 
- ou apenas um único handle válido

Montando o dicionários de handles

- numerar sequencialmente os "handle.invalid" de did's diferentes
- criar um "handle.missing" quando is.na(handle)
- remover o ".bsky.social" de quem tiver
- colar os handles dos did's que mudaram de handle pelo meio do caminho

```{r}
#| label: monta-dicionario-handles

dicionario.handles <- 
  did.handle.unicos |> 
  group_by(did) |> 
  filter(n() == 1 | n() > 1 & !is.na(handle)) |> 
  ungroup() |> 
  mutate(
    handle = 
      if_else(
        is.na(handle), 
        "handle.missing", 
        handle
      )
  ) |> 
  arrange(handle) |> 
  group_by(handle) |> # só missing e invalid é que sejam juntados 
  mutate(
    handle =
      if_else(
        n() == 1 + 0 * nchar(handle),
        handle |> 
          str_remove(fixed(".bsky.social")),
        handle |> 
          str_c(".", seq(n()))
      )
  )
dicionario.handles |> write_rds("Dados/dicionario-handles.rds")
```

Temos `r sum(str_sub(dicionario.handles$handle, 1, 14) == "handle.missing")` did's sem handle.  Vale a pena tentar puxar os dados desses did's com a 
`r dicionario.handles$did[12]`

`r which(str_sub(dicionario.handles$handle, 1, 14) == "handle.missing")`

## Se precisar, carregando do disco o que já foi processado

Só executa se avaliado manualmente

```{r}
#| label: recarrega-dados
#| eval: false

# likes
likes <- read_rds("Dados/likes.rds")
liker.actors <- read_rds("Dados/liker-actors.rds")
liked.actors <- read_rds("Dados/liked-actors.rds")
likes.actors <- read_rds("Dados/likes-actors.rds")
all.actors.likes <- read_rds("Dados/all-actors-likes.rds")
matriz.likes <- read_rds("Dados/matriz-likes.rds")
# skeets
skeets.longo <- read_rds("Dados/skeets-longo.rds")
matriz.skeets <- read_rds("Dados/matriz-skeets.rds")
all.actors.likes <- read_rds("Dados/all-actors-likes.rds")
skeets.separados <- 
  list(
    quote = read_rds("Dados/skeets-quote.rds"),
    reply = read_rds("Dados/skeets-reply.rds"),
    skeet = read_rds("Dados/skeets-iniciais.rds")
  )
```
