==> 03-metodol.qmd <==
# Metodologia

A extração de informações consiste na extração automática de informações estruturadas como entidades, relações entre entidades e atributos descrevendo entidades de fontes desordenadas, que permite buscas mais aprofundadas do que as possíveis por buscas de palavras-chave [@Sarawagi2008]. Para a extração de informações disponíveis *on-line*, esses dados podem ser acessados através de uma interface de programação de aplicações (API) própria de cada plataforma digital, porém grandes fontes como Facebook, Google e X têm recentemente limitado ou restringido o acesso gratuito às suas APIs em virtude do valor comercial dessas informações, dificultando a obtenção de dados reais para fins acadêmicos [@Batrinca2014]. Dentre as redes sociais mais relevantes atualmente, o Bluesky se destaca como uma das únicas que ainda permite acesso irrestrito e gratuito à sua API até o momento da redação deste trabalho, desde que se utilize de uma autenticação única atrelada a uma conta na plataforma.
 
Feita a coleta dos dados dessa plataforma, é possível empregar técnicas de análise de redes sociais para extrair *insights* acionáveis de um grande volume de informações desordenadas. Para a detecção de comunidades, @Aggarwal2011 sugere a representação através de grafos, que @Gomes2022 descreve como abstrações matemáticas que modelam relações simétricas dois a dois entre determinados elementos de um conjunto, tal que um grafo $G = \{V, E\}$ é composto de um conjunto $V$ de vértices e um conjunto $E$ de arestas, cada uma definida por um par não-ordenado de vértices. Chama-se de indicência a propriedade de uma aresta estar conectada a um vértice e de adjacência a propriedade de dois vértices estarem conectados por intermédio de uma aresta, sendo estes vértices adjacentes também chamados de "vizinhos" [@Gomes2022]. Esses grupos de vértices são então definidos como comunidades distintas se suas conexões internas são mais fortes do que com o resto da rede [@Aggarwal2011].

A medida de influência social na análise de redes sociais, de acordo com @Aggarwal2011, parte do princípio que as pessoas tendem a seguir ou adotar os comportamentos de seus amigos ou "vizinhos" (vértices adjacentes), fenômeno conhecido como homofilia. @Freeman1978, similarmente, aponta que alguém numa rede social em uma posição que permita contato direto com muitas outras pessoas pode vir a ser visto como um importante canal de informação e um ponto focal das comunicações. Essa influência pode ser entendida como um efeito direcional de um vértice $v_1$ para outro vértice $v_2$, relacionado com a força da aresta de $v_1$ para $v_2$, ao mesmo tempo em que a semelhança nas "vizinhanças" de $v_1$ e $v_2$ representam a força da aresta que os conecta, com muitos vizinhos em comum representando um laço forte e vice-versa [@Aggarwal2011]. 

As medidas de centralidade podem ser agrupadas em duas categorias: medidas radiais e medidas mediais, em que as medidas radiais avaliam caminhadas aleatórias que começam ou terminam em um determinado vértice e as medidas mediais avaliam caminhadas aleatórias que passam por um determinado vértice [@Aggarwal2011]. Dentre essas, as mais relevantes para o presente trabalho são as medidas de centralidade de grau, centralidade de proximidade e centralidade de intermediação.

A medida de centralidade de grau é a medida de centralidade radial mais simples, contando o número de arestas que originam de um vértice $i$ tal que $c_i^{DEG}=deg(i)$, em que $deg(i)$ representa o grau do vértice e $c_i^{DEG}$ a centralidade de grau [@Aggarwal2011]. Pode-se especular, portanto, que vértices com uma centralidade de grau relativamente alta têm uma visibilidade ampliada ou um alto potencial de atividade em comunicação dentro de uma rede social, enquanto vértices com um baixo grau são vistos como periféricos e de pouco envolvimento nos processos comunicativos [@Freeman1978].

Outro grupo de medidas radiais são as de centralidade de proximidade. A medida mais popular nesse grupo é a centralidade de proximidade de Freeman ($c_i^{CLO}$), que mede a centralidade de um vértice $i$ com base na média das distâncias mais curtas para todos os outros vértices no grafo, definida por $c_i^{CLO}=e^T_iS\mathbf1$, na qual $S$ é a matriz que contém o caminho de menor distância entre os vértices $i$ e $j$ e 1 é um vetor ou matriz de uns [@Aggarwal2011].

A centralidade de intermediação de Freeman ($c_i^{BET}$), medida de centralidade medial mais conhecida, mede o quanto um vértice $i$ está presente nos caminhos mais curtos da rede e é dada por $c_i^{BET}=\sum_{j,k}\frac{b_{jik}}{b_{jk}}$, onde $b_{jk}$ é o número de caminhos mais curtos entre os vértices $j$ e $k$ e $b_{jik}$ é o número de caminhos mais curtos entre $j$ e $k$ que passem pelo vértice $i$ [@Aggarwal2011]. Essa medida se baseia na suposição de que uma pessoa que está estrategicamente posicionada entre duas comunidades possui um potencial de controle sob as comunicações destas e, portanto, ocupa uma posição central [@Freeman1978].



## Tecnologias Utilizadas



### Ambiente Computacional Estatístico R

O ambiente computacional estatístico R, sua linguagem e de programação e suas bibliotecas têm se consolidado como uma das principais ferramentas da ciência de dados contemporânea, oferecendo recursos avançados para manuseio de dados, análise estatística, modelagem e visualização de informações. De acordo com @Wickham2017, o R é "uma linguagem poderosa e flexível, projetada para análise estatística e visualização de dados, tornando-se uma das principais ferramentas da ciência de dados moderna”. No presente trabalho, o R é utilizado tanto para realizar a mineração, tratamento e formatação de dados provenientes da rede social Bluesky, quanto para documentação do presente trabalho. 

## Fonte dos Dados

O Bluesky é uma rede social de API aberta construída a partir do Protocolo AT, uma fundação descentralizada para redes sociais públicas. Lançada ao público oficialmente em 2024, uma das propostas da rede social que justificam sua escolha para objeto deste estudo é o acesso aberto à sua arquitetura. A abertura do sistema proporciona a interação de seus conjuntos de dados para a moderação de conteúdos de mídia social e gerenciamento de comunidades (tanto partindo de usuários como partindo de pesquisadores). [@Kleppmann2024] Dessa forma, considerando sua arquitetura aberta e a disponibilidade ampla de dados para análise, a plataforma Bluesky foi selecionada como fonte provedora dos dados utilizados neste estudo.



### Mineração de Dados

Em seu artigo sobre abordagens quantitativas para a comparação de padrões comunicativos no Twitter (atual X), @Bruns2012 utilizam uma API de busca através da plataforma *yourTwapperkeeper* para extrair um conjunto de dados de postagens (*tweets*) que contenham palavras-chave ou *hashtags* determinadas pelo operador. No Bluesky, um processo equivalente pode ser realizado através do pacote '{atrrr}' do R para acessar a API da rede e buscar postagens e informações de usuários dentro de um critério estabelecido.

Em nosso estudo, a palavra-chave escolhida para a extração das postagens foi “CBLOL”, por se tratar do principal identificador do cenário competitivo brasileiro de League of Legends na plataforma. O uso do termo permite delimitar o recorte nacional das interações analisadas, uma vez que concentra a maior parte das discussões, transmissões e conteúdos produzidos sobre o campeonato. Além disso, durante o período de suspensão do X, os perfis oficiais do CBLOL utilizaram o Bluesky para transmitir informações e comentar as finais do Worlds, o que reforça a relevância do termo para capturar a atividade da comunidade.

Dessa forma, realizou-se a mineração de postagens (chamadas de skeets no Bluesky) contendo a palavra-chave “CBLOL”, bem como de suas interações, ao longo dos seguintes intervalos temporais:

- as 4 semanas anteriores à suspensão (03/08/2024 – 30/08/2024);
- as 5 semanas e 3 dias em que o X esteve indisponível (31/08/2024 – 08/10/2024);
- as 4 semanas posteriores ao retorno da rede (09/10/2024 – 05/11/2024).

Todo o fluxo de mineração está detalhado no documento 1-extracao.qmd, o qual será anexado como material suplementar ao final deste trabalho.

### Tratamento dos dados

Inicialmente, os dados foram divididos em dois grandes conjuntos: dados de likes e dados de skeets (postagens). Essa separação foi necessária porque cada conjunto contém estruturas distintas e cumpre funções analíticas específicas dentro do estudo. Enquanto os likes representam relações direcionadas entre atores — permitindo a construção de grafos e métricas de rede — os skeets fornecem informações temporais sobre a produção de conteúdo, essenciais para análises descritivas e temporais.

Após essa separação, realizou-se a padronização da identificação dos usuários. Cada ator presente nos dados foi associado ao seu respectivo identificador único (did), substituindo nomes e rótulos inconsistentes. Essa harmonização foi fundamental para evitar duplicidades e assegurar que as interações fossem corretamente atribuídas aos mesmos indivíduos ao longo dos diferentes arquivos.

Em seguida, foram criadas listas estruturadas de atores envolvendo relações de likes, tanto para quem dá likes quanto para quem os recebe. Essa etapa permitiu mapear todos os actors presentes na rede e identificar quais deles estavam ativos nas interações. A partir dessas listas, foi possível montar a matriz de likes dados, na qual cada linha representa uma interação do tipo “ator A curtiu ator B”, agregando o número de ocorrências dessa relação no banco de dados.

Todo o fluxo de tratamento — incluindo leitura dos arquivos, padronização de identificadores, criação de listas intermediárias e construção das matrizes — está detalhado no documento 2-tratamento-dados.qmd, o qual será anexado como material suplementar ao final deste trabalho. Esse arquivo contém o passo a passo completo das operações realizadas e serve como documentação técnica da preparação dos dados.

### Análise de Dados em Rede

No presente trabalho, utilizou-se da quantidade de curtidas (*likes*) em postagens (*skeets*) com o termo CBLOL como métrica principal para a propagação de opiniões. Essa escolha é baseada na premissa que curtidas representam um engajamento ativo e singular e que demonstram concordância ou aprovação em relação a um conteúdo pontual, enquanto a métrica de seguidores é indicativa somente de alcance potencial sem garantir engajamento ou influência real. Conforme demonstrado por @Moon2022, o número de seguidores não é sempre proporcional ao engajamento para campanhas de marketing de influência, com uma uma curva de formato U inverso sendo mais representativa da relação entre número de seguidores e engajamento, onde o engajamento é mais alto para influenciadores com um número moderado de seguidores. Adicionalmente, estudos revelam que usuários de redes sociais tendem a engajar mais (em termos de curtidas e *reposts*) conteúdos de contas que seguem menos pessoas, interpretando-as como mais autônomas [@Valsesia2020]. Ademais, embora as curtidas exijam menos esforço de interação do que os comentários [@Moon2022], elas se mostram superiores aos seguidores como métrica de interesse ativo. O número de seguidores frequentemente representa uma audiência passiva, visto que o volume de curtidas em perfis tende a ser significativamente mais baixo [@Avelino2020]. A união destes fatores, portanto, justifica as curtidas como sendo a métrica escolhida como mais apropriada para para mensurar a resposta efetiva do público e a dinâmica de disseminação na rede.

Todo o fluxo de metricas de redes está detalhado no documento 4-metricas-rede.qmd, o qual será anexado como material suplementar ao final deste trabalho.


### Análise de Dados em Rede - Visualizações 

No quinto arquivo de análise foram construídos e visualizados os grafos direcionados de interações para cada um dos 14 períodos definidos ao longo do tempo. A partir das matrizes de likes agregadas por período, geraram-se, para cada intervalo, objetos de grafo em que os nós representam atores e as arestas representam relações de engajamento por meio de likes.

Todo o fluxo de plotagem de grafos está detalhado no documento 5-grafos-likes.qmd, o qual será anexado como material suplementar ao final deste trabalho.
