==> 3-critica-likes.qmd <==
---
title: "Crítica dos dados de likes"
format: 
  html:
    toc: true
    toc-depth: 5
    toc-location: left
    embed-resources: true
execute: 
  cache: true
---

```{r}
#| label: setup
#| warning: false

library(tidyverse)
`%!in%` <- negate(`%in%`)
options(scipen = 99)
```

## Carregando dados

```{r}
#| label: carrega-likes

likes <- read_rds("Dados/likes.rds")
liker.actors <- read_rds("Dados/liker-actors.rds")
liked.actors <- read_rds("Dados/liked-actors.rds")
likes.actors <- read_rds("Dados/likes-actors.rds")
```


## Criticando lista de likers

```{r}
liker.actors |> 
  group_by(from.handle == "handle.invalid") |> 
  summarise(quantos = n())
```

`r sum(liker.actors$from.handle == "handle.invalid")` actors não têm mais handles válidos, então esses actors serão agrupados erradamente como um único actor se agrupar por handle.

```{r}
likes.liker.liked <- 
  tibble(
    actor = 
      c(liker.actors$from.actor, liked.actors) |> 
      unique() |> 
      sort()
  ) |> 
  mutate(
    curtiu = actor %in% liker.actors$from.actor,
    foi.curtido = actor %in% liked.actors
  ) |> 
  with(table(curtiu, foi.curtido)) |> 
  addmargins()
likes.liker.liked
```

Dos `r likes.liker.liked[2,3]` actors que curtiram skeets com "cblol" e dos `r likes.liker.liked[3,2]` actors postaram skeets com "cblol" que foram curtidos, `r likes.liker.liked[2,2]` curtiram e foram curtidos e `r likes.liker.liked[2,1]` curtiram mas não foram curtidos e `r likes.liker.liked[1,2]` foram curtidos mas não curtiram.

## Vendo se falta algum actor

Lendo skeets

```{r}
#| label: confere-actors

skeets <- read_rds("Dados/skeets-curto.rds")
all.actors.skeets <- 
  c(
    skeets$did.autor,
    skeets$uri.reply |> 
      str_remove("^at://") |> 
      str_remove("/app.bsky.feed.post/.*$") 
  ) |> 
  unique() |> 
  sort()
likes.actors <- 
  c(liker.actors$from.actor, liked.actors) |> 
  unique() |> 
  sort()
sum(all.actors.skeets %!in% likes.actors)
sum(likes.actors %!in% all.actors.skeets)
skeets$uri.reply |> 
  str_remove("^at://") |> 
  str_remove("/app.bsky.feed.post/.*$") |> 
  unique() |> 
  setdiff(likes.actors) |> 
  length()
```

Há `r skeets$uri.reply |> str_remove("^at://") |> str_remove("/app.bsky.feed.post/.*$") |> unique() |> setdiff(likes.actors) |> length()` autores de skeets sem "cblol" que estão nos skeets não estão nos likes porque eceberam replies com "cblol" curtidos 

## Verificando a matriz de likes

Carregando

```{r}
#| label: carrega-matriz-likes

matriz_likes <- read_rds("Dados/matriz-likes.rds")
```

Vendo como está para um dia específico (a final do campeonato, com muita atividade)

```{r}
matriz.likes.20240907 <- 
  matriz_likes |> 
  filter(date(data.skeet) == ymd("2024-09-07")) |> 
  count(
    actor.influencer.num, 
    actor.influenced.num, 
    name = "quantos"
  )
matriz.likes.20240907
```

Parece ok

Testando algumas medidas de centralidade em função dos likes

```{r}
all.actors.likes <- read_rds("Dados/all-actors-likes.rds")

matriz.likes.20240907.saindo <- 
  matriz.likes.20240907 |> 
  group_by(actor.influencer.num) |> 
  summarise(quantos = sum(quantos)) |> 
  left_join(all.actors.likes, by = c("actor.influencer.num" = "num")) |> 
  rename(
    actor.influencer = did,
    arestas.saindo = quantos
  )
matriz.likes.20240907.saindo

matriz.likes.20240907.entrando <- 
  matriz.likes.20240907 |> 
  group_by(actor.influenced.num) |> 
  summarise(quantos = sum(quantos)) |> 
  left_join(all.actors.likes, by = c("actor.influenced.num" = "num")) |> 
  rename(
    actor.influenced = did,
    arestas.entrando = quantos
  )
matriz.likes.20240907.entrando

matriz.likes.20240907.saldo <- 
  full_join(
    matriz.likes.20240907.saindo,
    matriz.likes.20240907.entrando,
    by = c("actor.influencer.num" = "actor.influenced.num"),
    keep = TRUE
  ) |> 
  mutate(
    across(
      c(arestas.saindo,arestas.entrando),
      \(x) if_else(x |> is.na(), 0, x)
    ),
    saldo.arestas = arestas.saindo - arestas.entrando,
    spread.arestas = arestas.saindo + arestas.entrando - abs(saldo.arestas),
  ) |>
  rename(
    actor = actor.influencer,
    actor.num = actor.influencer.num
  ) |> 
  select(
    actor, 
    actor.num, 
    arestas.saindo, 
    arestas.entrando,
    saldo.arestas, 
  )
matriz.likes.20240907.saldo |> 
  view()
```

Parece que tá funcionando
