==> 4-metricas-rede.qmd <==
---
title: "Metricas de rede"
format: 
  html:
    toc: true
    toc-depth: 5
    toc-location: left
    embed-resources: true
execute: 
  cache: refresh
---

```{r}
#| label: setup
#| warning: false

library(tidyverse)
`%!in%` <- negate(`%in%`)
options(scipen = 99)
```

## Carregando dados

```{r}
#| label: carrega-likes

# likes
likes <- read_rds("Dados/likes.rds")
liker.actors <- read_rds("Dados/liker-actors.rds")
liked.actors <- read_rds("Dados/liked-actors.rds")
likes.actors <- read_rds("Dados/likes-actors.rds")
all.actors.likes <- read_rds("Dados/all-actors-likes.rds")
matriz.likes <- read_rds("Dados/matriz-likes.rds")
```

## Verificando a matriz de likes para todos os dias

fazendo períodos

```{r}
#| label: períodos
 
inicio.bloqueio <- "30-08-2024" |> dmy()
(inicio.busca <- inicio.bloqueio - 4 * dweeks())

final.bloqueio <- "08-10-2024" |> dmy()
(final.busca <- final.bloqueio + 2 * ddays() + 4 * dweeks())

dias.periodo <- seq.Date(from = inicio.busca, to = final.busca, by = "day")
```
fazendo uma matriz para cada dia

```{r}
#| label: matriz-cada-dia

matriz.lista.dias <- 
  map(
    dias.periodo,
    ~matriz.likes |>
        filter(date(data.skeet) == .x) |> 
        count(
          actor.influencer.num, 
          actor.influenced.num, 
          name = "quantos"
        )
  )
  
names(matriz.lista.dias) <- dias.periodo |> format("%Y-%m-%d")

```

salvando

```{r}

walk(
  names(matriz.lista.dias),
  ~ write_rds(
      matriz.lista.dias[[.x]],
      paste0("Dados/matrizes-dia/", .x, "-likes.rds")
    )
)

```

criando um vetor das matrizes

```{r}

vetor.matrizes.dias.likes <- 
  list.files(
    path = "Dados/matrizes-dia",
    pattern = "\\.rds$",
    full.names = TRUE
  )
saveRDS(vetor.matrizes.dias.likes, "Dados/vetor-matrizes-dias-likes.rds")

```

não precisa rodar

quantidade de matrizes (98 dias)

```{r}
#| label: conta-arquivos
#| eval: false

n_arquivos <- length(list.files("Dados/matrizes-dia"))
n_arquivos

```


Vendo como está para um dia específico e tirando medidas de centralidade em função dos likes

Da matriz 1 a 27 - não há likes (ninguém curtiu ou foi curtido)

A partir do dia 28 - há likes

exemplo do dia [37] = 07-09-2024


```{r}
dia37 <- readRDS(vetor.matrizes.dias.likes[[37]])

# in-degree: Influencia sendo recebida (actor.influencer.num)
in_dia37 <- 
  dia37 |>
  group_by(actor.influencer.num) |>
  summarise(
    in_degree = sum(quantos),
    .groups = "drop"
  ) |>
  rename(actor.num = actor.influencer.num)

# out-degree: Influencia sendo dada (actor.influenced.num)
out_dia37 <- 
  dia37 |>
  group_by(actor.influenced.num) |>
  summarise(
    out_degree = sum(quantos),
    .groups = "drop"
  ) |>
  rename(actor.num = actor.influenced.num)

# junta in + out por ator
metricas_agregadas_dia37 <- 
  full_join(in_dia37, out_dia37, by = "actor.num") |>
  replace_na(list(in_degree = 0, out_degree = 0)) |>
  mutate(
    saldo  = in_degree - out_degree,
    spread = in_degree + out_degree - abs(saldo)
  ) |>
  left_join(
    all.actors.likes |> select(num, did, handle),
    by = c("actor.num" = "num")
  ) |>
  relocate(did, handle, .after = actor.num) |>
  arrange(desc(in_degree))

saveRDS(metricas_agregadas_dia37, "Dados/metricas_agregadas_dia37.rds")
```

```{r}
#TODOS OS DIAS

# matriz.lista.dias já tem names com datas no formato "YYYY-MM-DD"

metricas_todos_dias <- 
  imap_dfr(
    matriz.lista.dias,
    ~ {
      dia <- .x      # data.frame desse dia
      data_str <- .y # nome da lista, ex: "2024-09-07-likes"
      
      # in-degree: quem recebe likes
      in_dia <- 
        dia |>
        group_by(actor.influenced.num) |>
        summarise(
          in_degree = sum(quantos),
          .groups = "drop"
        ) |>
        rename(actor.num = actor.influenced.num)
      
      # out-degree: quem dá likes
      out_dia <- 
        dia |>
        group_by(actor.influencer.num) |>
        summarise(
          out_degree = sum(quantos),
          .groups = "drop"
        ) |>
        rename(actor.num = actor.influencer.num)
      
      # junta in + out para o dia
      full_join(in_dia, out_dia, by = "actor.num") |>
        replace_na(list(in_degree = 0, out_degree = 0)) |>
        mutate(
          saldo  = in_degree - out_degree,
          spread = in_degree + out_degree - abs(saldo),
          data   = ymd(data_str)
        ) |>
        left_join(
          all.actors.likes |> select(num, did, handle),
          by = c("actor.num" = "num")
        ) |>
        relocate(data, actor.num, did, handle)
    }
  )

saveRDS(metricas_todos_dias, "Dados/metricas-todos-dias.rds" )
```



# por 14 periodos

```{r}
#| label: separa-14-periodos

periodos_idx <- list(
  periodo1  = 2:8,
  periodo2  = 9:15,
  periodo3  = 16:22,
  periodo4  = 23:29,
  periodo5  = 30:36,
  periodo6  = 37:43,
  periodo7  = 44:50,
  periodo8  = 51:57,
  periodo9  = 58:64,
  periodo10 = 65:71,
  periodo11 = 72:78,
  periodo12 = 79:85,
  periodo13 = 86:92,
  periodo14 = 93:98
)
```

```{r}
matriz_periodos <- imap(
  periodos_idx,
  ~ matriz.lista.dias[.x] |>
    bind_rows() |>
    group_by(actor.influencer.num, actor.influenced.num) |>
    summarise(
      quantos = sum(quantos),
      .groups = "drop"
    )
)

# nomes são "periodo1", "periodo2", "periodo3"...
names(matriz_periodos)
saveRDS(matriz_periodos, "Dados/matriz_periodo5.rds")

```

```{r}

matriz_p5 <- matriz_periodos$periodo5

# 2) in-degree: influência recebida (actor.influencer.num)
in_p5 <- 
  matriz_p5 |>
  dplyr::group_by(actor.influencer.num) |>
  dplyr::summarise(
    in_degree = sum(quantos),
    .groups   = "drop"
  ) |>
  dplyr::rename(actor.num = actor.influencer.num)

# 3) out-degree: influência dada (actor.influenced.num)
out_p5 <- 
  matriz_p5 |>
  dplyr::group_by(actor.influenced.num) |>
  dplyr::summarise(
    out_degree = sum(quantos),
    .groups    = "drop"
  ) |>
  dplyr::rename(actor.num = actor.influenced.num)

# 4) junta in + out por ator, calcula saldo e spread, adiciona handle
metricas_agregadas_periodo5 <- 
  dplyr::full_join(in_p5, out_p5, by = "actor.num") |>
  tidyr::replace_na(list(in_degree = 0, out_degree = 0)) |>
  dplyr::mutate(
    saldo  = in_degree - out_degree,
    spread = in_degree + out_degree - abs(saldo)
  ) |>
  dplyr::left_join(
    all.actors.likes |> dplyr::select(num, did, handle),
    by = c("actor.num" = "num")
  ) |>
  dplyr::relocate(did, handle, .after = actor.num) |>
  dplyr::arrange(dplyr::desc(in_degree))

# 5) salva o objeto para uso futuro
saveRDS(metricas_agregadas_periodo5, "Dados/metricas_agregadas_periodo5.rds")
```

# por 3 periodos
```{r}
#| label: separa-tres-periodos

periodos3_idx <- list(
  periodo1 = 1:32,
  periodo2 = 33:65,
  periodo3 = 66:98
)
```

```{r}

matriz_3_periodos <- imap(
  periodos_idx,
  ~ matriz.lista.dias[.x] |>
    bind_rows() |>
    group_by(actor.influencer.num, actor.influenced.num) |>
    summarise(
      quantos = sum(quantos),
      .groups = "drop"
    )
)

# nomes são "periodo1", "periodo2", "periodo3"
names(matriz_3_periodos)


```
