==> 5-grafos-likes.qmd <==
---
title: "5-grafos-likes"
format: 
  html:
    toc: true
    toc-depth: 5
    toc-location: left
    embed-resources: true
execute: 
  cache: refresh
---

```{r}
#| label: setup
#| warning: false

library(tidyverse)
library(tidygraph)
library(ggraph)
library(igraph)
library(viridis)
library(purrr)
library(patchwork)
```

```{r}
#| label: carrega-metricas

all.actors.likes <- read_rds("Dados/all-actors-likes.rds")
matriz.likes <- read_rds("Dados/matriz-likes.rds")
vetor.matrizes.dias.likes <- readRDS("Dados/vetor-matrizes-dias-likes.rds")
dia37 <- readRDS(vetor.matrizes.dias.likes[[37]])
metricas_agregadas_dia37 <- read_rds("Dados/metricas_agregadas_dia37.rds")

#metricas_todos_dias <- ("Dados/metricas-todos-dias.rds")
```


```{r}
## 5.2 e 5.3. Construindo o grafo do dia 37

# edges: quem curte (from) → quem recebe like (to)
edges_dia37 <- 
  dia37 |>
  rename(
    from   = actor.influencer.num,
    to     = actor.influenced.num,
    weight = quantos
  ) |>
  mutate(
    from = as.character(from),
    to   = as.character(to)
  )

# nodes: todos os IDs que aparecem em from ou to
nodes_dia37 <- 
  tibble(name = unique(c(edges_dia37$from, edges_dia37$to))) |>
  left_join(
    all.actors.likes |>
      mutate(num = as.character(num)) |>
      select(name = num, handle),
    by = "name"
  )

# construindo o grafo com tidygraph
grafo_dia37 <- 
  tbl_graph(
    nodes = nodes_dia37,
    edges = edges_dia37,
    directed = TRUE
  )

```



```{r}
grafo_dia37 <- 
  grafo_dia37 |>
  activate(nodes) |>
  left_join(
    metricas_agregadas_dia37 |> 
      mutate(actor.num = as.character(actor.num)),
    by = c("name" = "actor.num")
  ) |>
  mutate(
    grau_in    = centrality_degree(mode = "in",  loops = FALSE),
    grau_out   = centrality_degree(mode = "out", loops = FALSE),
    betweenness = centrality_betweenness(),
    eigen       = centrality_eigen()
  )
```

Aí você passa a ter tudo junto:

in_degree / out_degree (baseados em soma de likes)

grau_in / grau_out (baseados em número de conexões diferentes)


```{r}
#as operações abaixo devem agir sobre os nós do grafo
metricas_dia37 <-
grafo_dia37 |>
activate(nodes) |>
as_tibble()

metricas_dia37 |> arrange(desc(grau_in)) |> head(20) |>
write_rds("Dados/metricas_dia37.rds")

```


mantendo quem recebeu 5+ curtidas por dia

```{r}

grafo_filtrado <-
grafo_dia37 |>
activate(nodes) |>
filter(grau_in >= 100)

grafo_filtrado
```


```{r}
# selecionar top 30 nós por grau_in (influencia entrando)

top_nos <- 
  grafo_dia37 |>
  activate(nodes) |>
  filter(rank(-grau_in) <= 30)


set.seed(123)

top_influenced <- ggraph(top_nos, layout = "fr") +
  
  # Arestas mais discretas
  geom_edge_link(
    color = "grey60",
    alpha = 0.8,
    show.legend = FALSE
  ) +
  
  # Nós coloridos pelo grau_out
  geom_node_point(
    aes(size = grau_in, color = grau_in),
    alpha = 0.8
  ) +
  
  # Labels com repel e cor branca
  geom_node_text(
    aes(label = handle.x),
    repel = TRUE,
    size = 3,
    color = "white",
    bg.color = "black",     # fundo do label
    bg.r = 0.15             # borda arredondada
  ) +
  # Escalas bonitas
  scale_color_viridis_c(
    name = "quantidade likes",
    option = "plasma"
  ) +
  scale_size_continuous(
    name = "quantidade likes",
    range = c(4, 14)
  ) +
  theme_void() +
  theme(
    plot.background = element_rect(fill = "white", color = NA),
    legend.position = "right",
    plot.title = element_text(color = "black", size = 16, face = "bold")
  ) +
  ggtitle("Rede de likes – top 30 atores que mais dão curtida | dia 07/09/2024")

# exportando imagem ----
ggplot2::ggsave(
  filename = "img/top30in.png",
  plot = top_influenced,
  width = 10,
  height = 7,
  dpi = 300
)

```


```{r}
# selecionar top 30 nós por grau_out (influencia saindo)

top_influencer <-
  grafo_dia37 |>
  activate(nodes) |>
  arrange(desc(grau_out)) |>
  slice(1:30)

set.seed(123)

top_influencer <- ggraph(top_influencer, layout = "fr") +
  
  # Arestas mais discretas
  geom_edge_link(
    color = "grey60",
    alpha = 0.8,
    show.legend = FALSE
  ) +
  
  # Nós coloridos pelo grau_out
  geom_node_point(
    aes(size = grau_out, color = grau_out),
    alpha = 0.8
  ) +
  
  # Labels com repel e cor branca
  geom_node_text(
    aes(label = handle.x),
    repel = TRUE,
    size = 3,
    color = "white",
    bg.color = "black",     # fundo do label
    bg.r = 0.15             # borda arredondada
  ) +
  
  # Escalas bonitas
  scale_color_viridis_c(
    name = "quantidade likes",
    option = "plasma"
  ) +
  
  scale_size_continuous(
    name = "quantidade likes",
    range = c(4, 14)
  ) +
  
  # Tema dark para destacar nós coloridos
  theme_void() +
  theme(
    plot.background = element_rect(fill = "white", color = NA),
    legend.position = "right",
    plot.title = element_text(color = "black", size = 16, face = "bold")
  ) +
  
  ggtitle("Rede de likes – top 30 atores que mais recebem curtida | dia 07/09/2024")

# exportando imagem ----
ggplot2::ggsave(
  filename = "img/top30out.png",
  plot = top_influencer,
  width = 10,
  height = 7,
  dpi = 300
)

```

# TODOS os dias

```{r}
# dicionário de atores (ajuste o nome do arquivo se for outro)
all.actors.likes <- read_rds("Dados/all-actors-likes.rds")

# vetor com os caminhos das matrizes por dia
vetor.matrizes.dias.likes <- read_rds("Dados/vetor-matrizes-dias-likes.rds")

```


2️⃣ Função que monta o grafo de um dia

Ela recebe o índice do vetor (idx) e devolve um tbl_graph.

```{r}

monta_grafo_dia <- function(idx) {
  # 1) carregar matriz do dia
  dia <- readRDS(vetor.matrizes.dias.likes[[idx]])
  
  # 2) edges: from influencer to influenced
  edges <- 
    dia |>
    rename(
      from   = actor.influencer.num,
      to     = actor.influenced.num,
      weight = quantos
    ) |>
    mutate(
      from = as.character(from),
      to   = as.character(to)
    )
  
  # 3) nodes: IDs únicos + handle
  nodes <- 
    tibble(name = unique(c(edges$from, edges$to))) |>
    left_join(
      all.actors.likes |>
        mutate(num = as.character(num)) |>
        select(name = num, handle),
      by = "name"
    )
  
  # 4) grafo com métricas básicas
  grafo <- 
    tbl_graph(
      nodes = nodes,
      edges = edges,
      directed = TRUE
    ) |>
    activate(nodes) |>
    mutate(
      grau_in  = centrality_degree(mode = "in"),
      grau_out = centrality_degree(mode = "out")
    )
}
```



# INFLUENCER

3️⃣ Função que gera o gráfico (ggraph) de um dia

```{r}

plota_rede_dia <- function(idx) {
  g <- 
    monta_grafo_dia(idx) |>
    activate(nodes) |>
    filter(grau_out >= 100)   # filtro opcional
  
  # extrair a data a partir do nome do arquivo (ex: "2024-09-07-likes.rds")
  data_lab <- basename(vetor.matrizes.dias.likes[[idx]]) |>
    stringr::str_remove("-likes.rds")
  
  set.seed(123)
  
  p <- 
    ggraph(g, layout = "fr") +
    geom_edge_link(
      aes(alpha = weight),
      color = "black",
      show.legend = FALSE
    ) +
    geom_node_point(
      aes(size = grau_out, color = grau_out),
      alpha = 0.8
    ) +
      geom_node_text(
    aes(label = handle),
    repel     = TRUE,
    size      = 2.5,
    color     = "black",
    bg.color  = "white",
    bg.r      = 0.15,
  ) +
    geom_node_text(
      aes(label = handle),
      repel = TRUE,
      size = 2.5,
      color = "black"
    ) +
    scale_color_viridis_c(
      option = "plasma",
      name   = "quantidade likes"
    ) +
    scale_size_continuous(
      range = c(3, 10),
      name  = "quantidade likes"
    ) +
    theme_void() +
    theme(
      plot.background = element_rect(fill = "white", color = NA),
      legend.position = "right",
      plot.title = element_text(color = "black", size = 12, face = "bold")
    ) +
    ggtitle(paste("Dia", data_lab))
  
  p
}
```

4️⃣ Gerar gráficos para todos os dias de uma vez

Aqui você cria uma lista de ggplots, um por dia.

```{r}

indices <- seq_along(vetor.matrizes.dias.likes)

plots_todos <- map(indices, plota_rede_dia)

```


5️⃣ SALVAR uma imagem por dia

Bom se você quiser escolher manualmente quais incluir no Quarto.

```{r}
dir.create("img/GRAFOS-OUT", showWarnings = FALSE)

walk2(
  plots_todos,
  indices,
  ~ ggsave(
      filename = file.path("img/GRAFOS-OUT", paste0("rede-dia-", .y, ".png")),
      plot     = .x,
      width    = 8,
      height   = 6,
      dpi      = 300
    )
)

```



# INFLUENCED

3️⃣ Função que gera o gráfico (ggraph) de um dia

```{r}

plota_rede_dia <- function(idx) {
  g <- 
    monta_grafo_dia(idx) |>
    activate(nodes) |>
    filter(grau_in >= 100)   # filtro opcional
  
  # extrair a data a partir do nome do arquivo (ex: "2024-09-07-likes.rds")
  data_lab <- basename(vetor.matrizes.dias.likes[[idx]]) |>
    stringr::str_remove("-likes.rds")
  
  set.seed(123)
  
  p <- 
  ggraph(g, layout = "fr") +
  geom_edge_link(
    aes(alpha = weight),
    color = "black",
    show.legend = FALSE
  ) +
  geom_node_point(
    aes(size = grau_in, color = grau_in),
    alpha = 0.8
  ) +
  geom_node_text(
    aes(label = handle),
    repel     = TRUE,
    size      = 2.5,
    color     = "black",
    bg.color  = "white",
    bg.r      = 0.15,
  ) +
  scale_color_viridis_c(
    option = "plasma",
    name   = "quantidade likes"
  ) +
  scale_size_continuous(
    range = c(3, 10),
    name  = "quantidade likes"
  ) +
  theme_void() +
  theme(
    plot.background = element_rect(fill = "white", color = NA),
    legend.position = "right",
    plot.title = element_text(color = "black", size = 12, face = "bold")
  ) +
  ggtitle(paste("Dia", data_lab))
}
```

4️⃣ Gerar gráficos para todos os dias de uma vez

Aqui você cria uma lista de ggplots, um por dia.

```{r}

indices2 <- seq_along(vetor.matrizes.dias.likes)

plots_todos2 <- map(indices, plota_rede_dia)

```


5️⃣ SALVAR uma imagem por dia

Bom se você quiser escolher manualmente quais incluir no Quarto.

```{r}
dir.create("img/GRAFOS-IN", showWarnings = FALSE)

walk2(
  plots_todos2,
  indices2,
  ~ ggsave(
      filename = file.path("img/GRAFOS-IN", paste0("rede-dia-", .y, ".png")),
      plot     = .x,
      width    = 8,
      height   = 6,
      dpi      = 300
    )
)

```


#por 3 Periodos


```{r}

grafos_3_periodos <- imap(
  matriz_3_periodos,
  ~ {
    edges <- 
      .x |>
      rename(
        from   = actor.influencer.num,
        to     = actor.influenced.num,
        weight = quantos
      ) |>
      mutate(
        from = as.character(from),
        to   = as.character(to)
      )
    
    nodes <- 
      tibble(name = unique(c(edges$from, edges$to))) |>
      left_join(
        all.actors.likes |>
          mutate(num = as.character(num)) |>
          select(name = num, handle),
        by = "name"
      )
    
    tbl_graph(
      nodes   = nodes,
      edges   = edges,
      directed = TRUE
    )
  }
)

#metricas

grafos_3_periodos <- imap(
  grafos_3_periodos,
  ~ .x |>
    activate(nodes) |>
    mutate(
      grau_in     = centrality_degree(mode = "in",  loops = FALSE),
      grau_out    = centrality_degree(mode = "out", loops = FALSE),
      betweenness = centrality_betweenness(),
      eigen       = centrality_eigen()
    )
)

```


##IN

```{r}

metricas_p1 <- grafos_3_periodos$periodo1 |> activate(nodes) |> as_tibble()
metricas_p1 |> arrange(desc(grau_in)) |> head(20)

```
```{r}
metricas_p2 <- grafos_3_periodos$periodo2 |> activate(nodes) |> as_tibble()
metricas_p2 |> arrange(desc(grau_in)) |> head(20)

```
```{r}
metricas_p3 <- grafos_3_periodos$periodo3 |> activate(nodes) |> as_tibble()
metricas_p3 |> arrange(desc(grau_in)) |> head(20)
```


plotando e salvando


```{r}
dir.create("img/GRAFOS-PERIODO-IN", showWarnings = FALSE)

limiar_in <- 100  # corte de likes recebidos

walk(
  names(grafos_3_periodos),
  ~ {
    periodo_nome <- .x
    g_full       <- grafos_periodos[[.x]]
    
    # pega o intervalo de dias desse período
    idx_periodo  <- periodos3_idx[[periodo_nome]]
    data_inicio  <- dias.periodo[min(idx_periodo)]
    data_fim     <- dias.periodo[max(idx_periodo)]
    
    data_lab <- paste0(
      format(data_inicio, "%d/%m/%Y"),
      " a ",
      format(data_fim, "%d/%m/%Y")
    )
    
    set.seed(123)  # layout reprodutível
    
    # 1) layout com o grafo COMPLETO
    layout_full <- create_layout(g_full, layout = "fr")
    
    # 2) FILTRA o layout: só nós com grau_in >= limiar_in
    layout_filt <- 
      layout_full |>
      dplyr::filter(!is.na(grau_in), grau_in >= limiar_in)
    
    # se quiser inspecionar quantos nós sobraram:
    # nrow(layout_filt)
    
    p <-
      ggraph(layout_filt) +
      geom_edge_link(
        aes(alpha = weight),
        color = "black",
        show.legend = FALSE
      ) +
      geom_node_point(
        aes(size = grau_in, color = grau_in),
        alpha = 0.8
      ) +
      geom_node_text(
        aes(label = handle),
        repel     = TRUE,
        size      = 2.5,
        color     = "black",
        bg.color  = "white",
        bg.r      = 0.15
      ) +
      scale_color_viridis_c(
        option = "plasma",
        name   = "quantidade likes"
      ) +
      scale_size_continuous(
        range = c(3, 10),
        name  = "quantidade likes"
      ) +
      theme_void() +
      theme(
        plot.background = element_rect(fill = "white", color = NA),
        legend.position = "none",
        plot.title      = element_text(color = "black", size = 12, face = "bold")
      ) +
      ggtitle(paste0(
        "Rede de likes – ", periodo_nome,
        "\n(", data_lab, ")",
        "\nNós com grau_in \u2265 ", limiar_in
      ))
    
    ggsave(
      filename = file.path("img/GRAFOS-PERIODO-IN",
                           paste0("grafo-", periodo_nome,
                                  "-in", limiar_in, ".png")),
      plot   = p,
      width  = 8,
      height = 6,
      dpi    = 300
    )
  }
)
```


##OUT

```{r}

metricas_p1 <- grafos_3_periodos$periodo1 |> activate(nodes) |> as_tibble()
metricas_p1 |> arrange(desc(grau_out)) |> head(20)

```
```{r}
metricas_p2 <- grafos_periodos$periodo2 |> activate(nodes) |> as_tibble()
metricas_p2 |> arrange(desc(grau_out)) |> head(20)

```
```{r}
metricas_p3 <- grafos_periodos$periodo3 |> activate(nodes) |> as_tibble()
metricas_p3 |> arrange(desc(grau_out)) |> head(20)
```


plotando e salvando


```{r}
dir.create("img/GRAFOS-PERIODO-OUT", showWarnings = FALSE)

limiar_in <- 100  # corte de likes recebidos

walk(
  names(grafos_3_periodos),
  ~ {
    periodo_nome <- .x
    g_full       <- grafos_periodos[[.x]]
    
    # pega o intervalo de dias desse período
    idx_periodo  <- periodos3_idx[[periodo_nome]]
    data_inicio  <- dias.periodo[min(idx_periodo)]
    data_fim     <- dias.periodo[max(idx_periodo)]
    
    data_lab <- paste0(
      format(data_inicio, "%d/%m/%Y"),
      " a ",
      format(data_fim, "%d/%m/%Y")
    )
    
    set.seed(123)  # layout reprodutível
    
    # 1) layout com o grafo COMPLETO
    layout_full <- create_layout(g_full, layout = "fr")
    
    # 2) FILTRA o layout: só nós com grau_in >= limiar_in
    layout_filt <- 
      layout_full |>
      dplyr::filter(!is.na(grau_out), grau_out >= limiar_in)
    
    p <-
      ggraph(layout_filt) +
      geom_edge_link(
        aes(alpha = weight),
        color = "black",
        show.legend = FALSE
      ) +
      geom_node_point(
        aes(size = grau_in, color = grau_in),
        alpha = 0.8
      ) +
      geom_node_text(
        aes(label = handle),
        repel     = TRUE,
        size      = 2.5,
        color     = "black",
        bg.color  = "white",
        bg.r      = 0.15
      ) +
      scale_color_viridis_c(
        option = "plasma",
        name   = "quantidade likes"
      ) +
      theme_void() +
      theme(
        plot.background = element_rect(fill = "white", color = NA),
        legend.position = "right",
        plot.title      = element_text(color = "black", size = 12, face = "bold")
      ) +
      ggtitle(paste0(
        "Rede de likes – ", periodo_nome,
        "\n(", data_lab, ")",
        "\nNós com grau_out \u2265 ", limiar_in
      ))
    
    ggsave(
      filename = file.path("img/GRAFOS-PERIODO-OUT",
      paste0("grafo-", periodo_nome,"-in", limiar_in, ".png")),
      plot   = p,
      width  = 8,
      height = 6,
      dpi    = 300
    )
  }
)
```

# 14 PERIODOS

```{r}

grafos_periodos <- imap(
  matriz_periodos,
  ~ {
    edges <- 
      .x |>
      dplyr::rename(
        from   = actor.influencer.num,
        to     = actor.influenced.num,
        weight = quantos
      ) |>
      dplyr::mutate(
        from = as.character(from),
        to   = as.character(to)
      )
    
    nodes <- 
      tibble::tibble(name = unique(c(edges$from, edges$to))) |>
      dplyr::left_join(
        all.actors.likes |>
          dplyr::mutate(num = as.character(num)) |>
          dplyr::select(name = num, handle),
        by = "name"
      )
    
    tidygraph::tbl_graph(
      nodes   = nodes,
      edges   = edges,
      directed = TRUE
    ) |>
      tidygraph::activate(nodes) |>
      dplyr::mutate(
        grau_in  = tidygraph::centrality_degree(mode = "in"),
        grau_out = tidygraph::centrality_degree(mode = "out")
      )
  }
)

names(grafos_periodos)
# agora deve ter "periodo1" ... "periodo15"

```


# OUT 14 PERIODOS

```{r}
metricas_periodos <- imap(
  grafos_periodos,
  ~ .x |> 
    activate(nodes) |> 
    as_tibble()
)
```

```{r}
top20_out_periodos <- imap(
  metricas_periodos,
  ~ .x |> arrange(desc(grau_out)) |> slice(1:20)
)
```


plotando e salvando

```{r}

dir.create("img/GRAFOS-14-PERIODO-OUT-500", showWarnings = FALSE)

limiar_out <- 500  # corte de likes dados

walk(
  names(grafos_periodos),
  ~ {
    periodo_nome <- .x
    g_full       <- grafos_periodos[[.x]]
    
    # pega o intervalo de dias desse período
    idx_periodo  <- periodos_idx[[periodo_nome]]
    data_inicio  <- dias.periodo[min(idx_periodo)]
    data_fim     <- dias.periodo[max(idx_periodo)]
    
    data_lab <- paste0(
      format(data_inicio, "%d/%m/%Y"),
      " a ",
      format(data_fim, "%d/%m/%Y")
    )
    
    # 🔸 1) filtra nós no GRAFO (não no layout)
    g_filt <- 
      g_full |>
      activate(nodes) |>
      filter(!is.na(grau_out), grau_out >= limiar_out)
    

    
    set.seed(123)  # layout reprodutível
    
    p <-
      ggraph(g_filt, layout = "fr") +
      geom_edge_link(
        aes(alpha = weight),
        color = "black",
        show.legend = FALSE
      ) +
      geom_node_point(
        aes(size = grau_in, color = grau_in),
        alpha = 0.8
      ) +
      geom_node_text(
        aes(label = handle),
        repel     = TRUE,
        size      = 2.5,
        color     = "black",
        bg.color  = "white",
        bg.r      = 0.15
      ) +
      scale_color_viridis_c(
        option = "plasma",
        name   = "quantidade likes"
      ) +
      theme_void() +
      theme(
        plot.background = element_rect(fill = "white", color = NA),
        legend.position = "right",
        plot.title      = element_text(color = "black", size = 12, face = "bold")
      ) +
      ggtitle(paste0(
        "Rede de likes – ", periodo_nome,
        "\n(", data_lab, ")",
        "\nNós com grau_out \u2265 ", limiar_out
      ))
    
    ggsave(
      filename = file.path(
        "img/GRAFOS-14-PERIODO-OUT-500",
        paste0("grafo-", periodo_nome, "-out", ".png")
      ),
      plot   = p,
      width  = 8,
      height = 6,
      dpi    = 300
    )
  }
)


```

```{r}

length(grafos_periodos)
length(periodos_idx)
```
```{r}

metricas_p5 <- grafos_periodos$periodo5 |> activate(nodes) |> as_tibble()
metricas_p5 |> arrange(desc(grau_out)) |> head(20) 
  
saveRDS(metricas_p5, "Dados/metricas_periodo5.rds")

```
